requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
litellm>=1.40.0
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
